Moved csv generator to it's own file.
Prevents the webview failing to load.
